revert: Move get_normalized_attributes utility in Utils Class
